requirements.txt
werkzeug==3.0.3
flask==3.0.3
SQLAlchemy==2.0.30
PyMySQL==1.1.0
Flask-WTF==1.2.1
Flask-Login==0.6.3
Flask_Migrate==4.0.7
waitress==3.0.0
python-dotenv==1.0.1